Added routes and controllers for the application.
